Code Maintenance - Nested Namespaces - Added 'const' to a lot of function params - Anonymous namespaces for cpp file only functions - 'static_cast' and 'reinterpret_cast' instead of c style casts - Made some functions static or const - Removed unnecessary includes - Deleted copy and move constructor for most classes

+++ Mixture/include/Mixture/Input/MouseCodes.hpp
@@ -5,7 +5,7 @@
 namespace Mixture 
 {
 
-	using MouseCode = uint16_t;
+	using MouseCode = uint8_t;
 
 	namespace Mouse 
 	{
